Update API documentation and refactor group and user management endpoints

+++ src/main/java/com/example/kadan/service/GroupService.java
@@ -48,7 +48,7 @@ public class GroupService {
     public GroupResponseDto getGroupById(UUID currentUser, UUID id) {
         Group group = groupRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Group not found"));
         if (!group.hasMember(currentUser)) {
-            throw new EntityNotFoundException("Group not found");
+            throw new EntityNotFoundException("User not found");
         }
 
         return GroupResponseDto.fromEntity(group);
@@ -58,7 +58,7 @@ public class GroupService {
     public GroupResponseDto updateGroup(UUID currentUser, UUID id, UpdateGroupDto groupDto) {
         Group group = groupRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Group not found"));
         if (!group.hasMember(currentUser)) {
-            throw new EntityNotFoundException("Group not found");
+            throw new EntityNotFoundException("User not found");
         }
 
         if (StringUtils.isNotBlank(groupDto.name())) group.setName(groupDto.name());
@@ -75,7 +75,7 @@ public class GroupService {
     public void addMemberToGroup(UUID currentUser, UUID groupId, UUID memberId) {
         Group group = groupRepository.findById(groupId).orElseThrow(() -> new EntityNotFoundException("Group not found"));
         if (!group.hasMember(currentUser)) {
-            throw new EntityNotFoundException("Group not found");
+            throw new EntityNotFoundException("User not found");
         }
 
         //Return success if member already part of group
@@ -91,6 +91,7 @@ public class GroupService {
     }
 
     //TODO Users other than owner in should be in pending state in group memeber repo until they accept.
+    //TODO members list can be null. should handle that case.
     @Transactional
     public CreateGroupResponseDto createGroup(UUID currentUser, GroupDto groupDto) {
         User groupCreatorUser = userRepository.findById(currentUser).orElseThrow(() -> new EntityNotFoundException("User not found"));
@@ -120,25 +121,36 @@ public class GroupService {
         return CreateGroupResponseDto.fromEntity(savedGroup);
     }
 
+    //TODO If owner is removed, transfer ownership to another member.
+    //TODO Prevent removing self if owner and other members exist. maybe possible.
     @Transactional
     public void removeMemberFromGroup(UUID currentUser, UUID groupId, UUID memberId) {
         Group group = groupRepository.findById(groupId).orElseThrow(() -> new EntityNotFoundException("Group not found"));
-        if (!group.hasMember(currentUser)) {
-            throw new EntityNotFoundException("Group not found");
+        if (!group.hasMember(currentUser) || !group.hasMember(memberId)) {
+            throw new EntityNotFoundException("User not found");
+        }
+        if (group.getMembers().size() == 1) {
+            throw new DataIntegrityViolationException("Cannot remove the only member of the group");
+        }
+
+        BalanceResponseDto response = balanceService.getBalances(group);
+        if (response.doesMemberHaveBalance(memberId)) {
+            throw new DataIntegrityViolationException("Cannot remove member with non-zero balances");
         }
 
         User memberToRemove = userRepository.findById(memberId).orElseThrow(() -> new EntityNotFoundException("User not found"));
         groupMemberRepository.deleteByGroupAndUser(group, memberToRemove);
     }
 
+    //TODO maybe consider only owners can delete groups.
     @Transactional
     public void deleteGroup(UUID currentUser, UUID groupId) {
         Group group = groupRepository.findById(groupId).orElseThrow(() -> new EntityNotFoundException("Group not found"));
         if (!group.hasMember(currentUser)) {
-            throw new EntityNotFoundException("Group not found");
+            throw new EntityNotFoundException("User not found");
         }
 
-        BalanceResponseDto groupBalances = balanceService.getBalances(currentUser, groupId);
+        BalanceResponseDto groupBalances = balanceService.getBalances(group);
         if (!groupBalances.balances().isEmpty()) {
             throw new DataIntegrityViolationException("Cannot delete group with non-zero balances");
         }
